Refactor storage in a PetPlugin.
